update paddlex ci (#2971)

+++ models/PaddleX/ci/ci_run.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
 
-SUITE=$1
-NEW_MODEL_INFO_FILE=$2
-MEM_SIZE=16
+MODE=$1
+MODEL_LIST_FILE=$2
 
+################################################### 可配置环境变量 #####################################################
+# MEM_SIZE: 显存大小，默认值16G，设置示例：export MEM_SIZE=16
+# DEVICE_TYPE: 设备类型，默认gpu，只支持小写，设置示例：export DEVICE_TYPE=gpu
+# DEVICE_NUM: 卡数，默认4卡，设置示例：export DEVICE_NUM=4
+# TEST_RANGE: 测试范围，默认为空，设置示例：export TEST_RANGE='inference'
 
-set -x
-if [[ $SUITE != 'PaddleX' ]];then
-    set -e
-else
-    failed_model_info=""
+# set -x
+if [[ $MODE == 'PaddleX' ]];then
+    failed_cmd_list=""
 fi
 
 
@@ -33,23 +35,31 @@ function func_parser_dataset_url(){
     echo ${tmp}
 }
 
+function get_device_list(){
+    gn=`expr $DEVICE_NUM - 1`
+    seq=`seq -s "," 0 $gn`
+    if [[ $suite_name == "PaddleTS" ]];then
+        seq=0
+    fi
+    echo ${DEVICE_TYPE}:$seq
+}
+
 # 运行命令并输出结果，PR级CI失败会重跑3次并异常退出，增量级和全量级会记录失败命令，最后打印失败的命令并异常退出
 function run_command(){
-    set +e
     command=$1
     module_name=$2
     time_stamp=$(date +"%Y-%m-%d %H:%M:%S")
     command="timeout 30m ${command}"
-    printf "\e[32m|%-20s| %-20s | %-50s | %-20s\n\e[0m" "[${time_stamp}]" ${module_name} "Run ${command}"
+    printf "\e[32m|%-20s| %-50s | %-20s\n\e[0m" "[${time_stamp}]" "${command}"
     eval $command
     last_status=${PIPESTATUS[0]}
-    if [[ $SUITE != 'PaddleX' ]];then
+    if [[ $MODE != 'PaddleX' ]];then
         n=1
         # Try 3 times to run command if it fails
         while [[ $last_status != 0 ]]; do
             sleep 10
             n=`expr $n + 1`
-            printf "\e[32m|%-20s| %-20s | %-50s | %-20s\n\e[0m" "[${time_stamp}]" ${module_name} "Retrying $n times with command: ${command}"
+            printf "\e[32m|%-20s| %-50s | %-20s\n\e[0m" "[${time_stamp}]" "${command}"
             eval $command
             last_status=${PIPESTATUS[0]}
             if [[ $n -eq 3 && $last_status != 0 ]]; then
@@ -59,11 +69,10 @@ function run_command(){
         done
     else
         if [[ $last_status != 0 ]];then
-            failed_model_info="$failed_model_info \n ${module_name} | command: ${command}"
+            failed_cmd_list="$failed_cmd_list \n ${module_name} | command: ${command}"
             echo "Run ${command} failed"
         fi
     fi
-    set -e
 }
 
 # 准备数据集并做数据校验
@@ -134,29 +143,34 @@ function run_models(){
         IFS=$'|'
         run_model_list=(${run_model})
         for mode in ${run_model_list[@]};do
-            set +e
             black_model=`eval echo '$'"${mode}_black_list"|grep "^${model_name}$"`
-            set -e
             if [[ ! -z $black_model ]];then
                 # 黑名单模型，不运行
                 echo "$model_name is in ${mode}_black_list, so skip it."
                 continue
             fi
-            # 适配导出模型时，需要指定输出路径
-            base_mode_cmd="${PYTHON_PATH} main.py -c ${config_path} -o Global.mode=${mode} -o Train.epochs_iters=${epochs_iters} -o Train.batch_size=${batch_size} -o Evaluate.weight_path=${evaluate_weight_path} -o Predict.model_dir=${inference_weight_path}"
-            if [[ $mode == "export" ]];then
-                model_export_output_path=${model_output_path}/export
-                mkdir -p $model_export_output_path
-                weight_dict[$model_name]="$model_export_output_path"
-                run_mode_cmd="${base_mode_cmd} -o Global.output=${model_export_output_path}"
-            else
-                run_mode_cmd="${base_mode_cmd} -o Global.output=${model_output_path}"
-            fi
-            run_command ${run_mode_cmd} ${module_name}
-            # 在predict模式下，再次验证官方模型预测
-            if [[ $mode == "predict" ]];then
-                offcial_model_predict_cmd="${PYTHON_PATH} main.py -c ${config_path} -o Global.mode=${mode} -o Predict.model_dir=None -o Global.output=${model_output_path}_offical_predict"
+            # TEST_RANGE 为空时为普通全流程测试
+            if [[ -z $TEST_RANGE ]];then
+                # 适配导出模型时，需要指定输出路径
+                device_info=$(get_device_list)
+                base_mode_cmd="${PYTHON_PATH} main.py -c ${config_path} -o Global.mode=${mode} -o Global.device=${device_info} -o Train.epochs_iters=${epochs_iters} -o Train.batch_size=${batch_size} -o Evaluate.weight_path=${evaluate_weight_path} -o Predict.model_dir=${inference_weight_path}"
+                if [[ $mode == "export" ]];then
+                    model_export_output_path=${model_output_path}/export
+                    mkdir -p $model_export_output_path
+                    weight_dict[$model_name]="$model_export_output_path"
+                    run_mode_cmd="${base_mode_cmd} -o Global.output=${model_export_output_path}"
+                else
+                    run_mode_cmd="${base_mode_cmd} -o Global.output=${model_output_path}"
+                fi
+                run_command ${run_mode_cmd} ${module_name}
+            # TEST_RANGE 为inference时，只测试官方模型预测
+            elif [[ $TEST_RANGE == "inference" && $mode == "predict" ]];then
+                offcial_model_predict_cmd="${PYTHON_PATH} main.py -c ${config_path} -o Global.mode=predict -o Predict.model_dir=None -o Global.output=${model_output_path}_offical_predict"
                 run_command ${offcial_model_predict_cmd} ${module_name}
+                continue
+            # TEST_RANGE 为其他非空值时，不做模型级别测试
+            else
+                continue
             fi
         done
         if [[ ! -z $black_model ]];then
@@ -167,13 +181,16 @@ function run_models(){
         if [[ $check_options == 'null' ]]; then
             continue
         fi
-        check_options_list=(${check_options})
-        for check_option in ${check_options_list[@]};do
-            # 运行产出检查脚本
-            checker_cmd="${PYTHON_PATH} ${BASE_PATH}/checker.py --check --$check_option --output ${model_output_path} --check_weights_items ${check_weights_items} --module_name ${module_name}"
-            run_command ${checker_cmd} ${module_name}
-        done
-        rm -rf ${model_output_path}/*[0-9]*
+        # runing_train为空时为普通全过程测试
+        if [[ -z $TEST_RANGE ]];then
+            check_options_list=(${check_options})
+            for check_option in ${check_options_list[@]};do
+                # 运行产出检查脚本
+                checker_cmd="${PYTHON_PATH} ${BASE_PATH}/checker.py --check --$check_option --output ${model_output_path} --check_weights_items ${check_weights_items} --module_name ${module_name}"
+                run_command ${checker_cmd} ${module_name}
+            done
+            rm -rf ${model_output_path}/*[0-9]*
+        fi
     done
     model_dict[$module_name]="$model_list"
 }
@@ -181,48 +198,67 @@ function run_models(){
 # 准备数据集和运行模型
 function prepare_and_run(){
     config_file_list=$1
-    echo $config_file_list
-    prepare_dataset
-    run_models "${config_file_list}"
+    if [[ ! -z $config_file_list ]];then
+        # runing_train为空时为普通全过程测试,非空时代表无训练测试，无需准备数据集
+        if [[ -z $TEST_RANGE ]];then
+            prepare_dataset
+        fi
+        run_models "${config_file_list}"
+    fi
 }
 
 
 #################################################### PaddleX CI ######################################################
 # PaddleX CI 区分为 PR 级、套件级、全量级和增量级
-# PR 级：运行 PaddleX_simplify_models.txt 中的重点模型列表
-# 套件级：运行 PaddleX_simplify_models.txt 中指定套件的部分模型
-# 全量级：根据 ci_info.txt 抓取 PaddleX 支持的所有模型并测试
-# 增量级：根据 ci_info.txt 和 传入的 changed_yaml.txt 抓取新增的模型，对新增模型进行增量测试
+# PR 级：运行 pr_list.txt 中的重点模型列表，示例 bash ci_run.sh
+# 套件级：运行 pr_list.txt 中指定套件的部分模型，示例 bash ci_run.sh PaddleClas
+# 全量级：根据 config.txt 抓取 PaddleX 支持的所有模型并测试，示例 bash ci_run.sh PaddleX
+# 增量级：根据 config.txt 和 传入的 changed_list.txt 抓取新增或修改的模型，对新增模型进行增量测试，示例 bash ci_run.sh PaddleX changed_list.txt
+
 
 # 指定 python
 PYTHON_PATH="python"
 # 获取当前脚本的绝对路径，获得基准目录
 BASE_PATH=$(cd "$(dirname $0)"; pwd)
 MODULE_OUTPUT_PATH=${BASE_PATH}/outputs
+CONFIG_FILE=${BASE_PATH}/config.txt
 pip config set global.index-url https://mirrors.bfsu.edu.cn/pypi/web/simple
 declare -A weight_dict
 declare -A model_dict
 
 # 安装paddlex，完成环境准备
-if [[ -z $SUITE ]]; then
-    install_deps_cmd="pip install -e . && paddlex --install -y"
-    modules_info_file=${BASE_PATH}/PaddleX_simplify_models.txt
-elif [[ $SUITE == "PaddleX" ]];then
-    install_deps_cmd="pip install -e . && paddlex --install -y"
-    modules_info_file=${BASE_PATH}/ci_info.txt
-else
-    install_deps_cmd="pip install -e . && paddlex --install --use_local_repos $SUITE"
-    modules_info_file=${BASE_PATH}/PaddleX_simplify_models.txt
+install_pdx_cmd="pip install -e ."
+eval $install_pdx_cmd
+
+
+if [[ -z $MEM_SIZE ]]; then
+    MEM_SIZE=16
+fi
+if [[ -z $DEVICE_TYPE ]]; then
+    DEVICE_TYPE='gpu'
+fi
+if [[ -z $DEVICE_NUM ]]; then
+    DEVICE_NUM=4
 fi
 
-eval ${install_deps_cmd}
+if [[ -z $MODE ]]; then
+    install_deps_cmd="paddlex --install -y"
+elif [[ $MODE == "PaddleX" ]];then
+    install_deps_cmd="paddlex --install -y"
+else
+    install_deps_cmd="paddlex --install --use_local_repos $MODE"
+fi
+
+# # 只测试产线推理无需安装套件库
+if [[ $TEST_RANGE != "pipeline" ]];then
+    eval ${install_deps_cmd}
+fi
 pip freeze > all_packages.txt
 
 
 #################################################### 模型级测试 ######################################################
 IFS=$' '
 black_list_file=${BASE_PATH}/black_list.txt
-set +e
 all_black_list=`cat ${black_list_file} | grep All: | awk -F : {'print$2'}`
 train_black_list=`cat ${black_list_file} | grep Train: | awk -F : {'print$2'}`
 train_black_list="$all_black_list
@@ -249,11 +285,10 @@ $export_black_list
 ##############pipeline_black_list###############
 $pipeline_black_list
 -----------------------------------------------------------------"
-set -e
 
 IFS='*'
-modules_info_list=($(cat ${modules_info_file}))
-all_module_names=`cat $modules_info_file | grep module_name | awk -F ':' {'print$2'}`
+modules_info_list=($(cat ${CONFIG_FILE}))
+all_module_names=`cat $CONFIG_FILE | grep module_name | awk -F ':' {'print$2'}`
 
 unset http_proxy https_proxy
 IFS=$' '
@@ -288,20 +323,18 @@ for modules_info in ${modules_info_list[@]}; do
             inference_model_dir=$(func_parser_value "${lines[line_num]}")
             line_num=`expr $line_num + 1`
             epochs_iters=$(func_parser_value "${lines[line_num]}")
-            if [[ $SUITE == "PaddleX" ]];then
-                if [[ ! -z $NEW_MODEL_INFO_FILE ]];then
-                    new_model_info=`cat $NEW_MODEL_INFO_FILE`
-                    new_model_module_names=`cat $NEW_MODEL_INFO_FILE | awk -F '/' {'print$3'} | sort -u`
+            if [[ $MODE == "PaddleX" ]];then
+                if [[ ! -z $MODEL_LIST_FILE ]];then
+                    new_model_info=`cat $MODEL_LIST_FILE`
+                    new_model_module_names=`cat $MODEL_LIST_FILE | awk -F '/' {'print$3'} | sort -u`
                     for new_module_info in ${new_model_module_names[@]};do
-                        set +e
                         module=`echo "${all_module_names[@]}" | grep $new_module_info`
-                        set -e
                         if [[ -z $module ]];then
                             echo "new module: $new_module_info is unsupported! Please contact with the developer or add new module info in ci_info.txt!"
                             exit 1
                         fi
                         if [[ $new_module_info == $module_name ]];then
-                            module_info=`cat $NEW_MODEL_INFO_FILE | grep $module_name | xargs -n1 -I {} echo config_path:{}`
+                            module_info=`cat $MODEL_LIST_FILE | grep $module_name | xargs -n1 -I {} echo config_path:{}`
                             echo $module_info
                             prepare_and_run "${module_info}"
                         fi
@@ -311,11 +344,11 @@ for modules_info in ${modules_info_list[@]}; do
                     prepare_and_run "${module_info}"
                 fi
                 continue
-            fi
-        elif [[ ! -z $module_info ]]; then
-            if [[ ! -z $SUITE && $SUITE == $suite_name ]];then
+            elif [[ $MODE == $suite_name ]];then
+                module_info=`cat ${BASE_PATH}/pr_list.txt | grep -v "^#" | grep paddlex/configs/${module_name}`
                 prepare_and_run "${module_info}"
-            elif [[ -z $SUITE ]];then
+            elif [[ -z $MODE ]];then
+                module_info=`cat ${BASE_PATH}/pr_list.txt | grep -v "^#" | grep paddlex/configs/${module_name}`
                 prepare_and_run "${module_info}"
             else
                 continue
@@ -339,6 +372,7 @@ function check_pipeline() {
 	mkdir -p $output_path
 	cd $output_path
     cmd="timeout 30m paddlex --pipeline ${pipeline} --input ${img}"
+	echo $cmd
 	eval $cmd
     last_status=${PIPESTATUS[0]}
     if [[ $last_status != 0 ]];then
@@ -349,9 +383,9 @@ function check_pipeline() {
 
 for pipeline_yaml in ${PIPELINE_YAML_LIST[@]};do
     pipeline_name=`cat paddlex/pipelines/${pipeline_yaml} | grep pipeline_name | awk {'print$2'}`
-    set +e
+    IFS=' '
     black_pipeline=`echo ${pipeline_black_list}|grep "^${pipeline_name}$"`
-    set -e
+    IFS=$'\n'
     if [[ ! -z $black_pipeline ]];then
         # 黑名单产线，不运行
         echo "$pipeline_name is in pipeline_black_list, so skip it."
@@ -361,7 +395,7 @@ for pipeline_yaml in ${PIPELINE_YAML_LIST[@]};do
     check_pipeline $pipeline_name "" "" $input
 done
 
-if [[ $SUITE == 'PaddleX' && ! -z $failed_model_info ]];then
-    echo $failed_model_info
+if [[ $MODE == 'PaddleX' && ! -z $failed_cmd_list ]];then
+    echo $failed_cmd_list
     exit 1
 fi
